+++ libgroup.cpp
@@ -428,7 +428,7 @@ LUA_FUNCTION(CheckWithSumEqual) {
 		max = min;
 	int32_t extraargs = lua_gettop(L) - 5;
 	card_vector cv(pduel->game_field->core.must_select_cards);
-	int32_t mcount = cv.size();
+	int32_t mcount = static_cast<int32_t>(cv.size());
 	const auto beginit = pduel->game_field->core.must_select_cards.begin();
 	const auto endit = pduel->game_field->core.must_select_cards.end();
 	for(auto& pcard : pgroup->container) {
@@ -466,7 +466,7 @@ LUA_FUNCTION(SelectWithSumEqual) {
 		pduel->game_field->core.select_cards.erase(it, pduel->game_field->core.select_cards.end());
 	}
 	card_vector cv(pduel->game_field->core.must_select_cards);
-	int32_t mcount = cv.size();
+	int32_t mcount = static_cast<int32_t>(cv.size());
 	cv.insert(cv.end(), pduel->game_field->core.select_cards.begin(), pduel->game_field->core.select_cards.end());
 	for(auto& pcard : cv)
 		pcard->sum_param = pduel->lua->get_operation_value(pcard, findex, extraargs);
@@ -493,7 +493,7 @@ LUA_FUNCTION(CheckWithSumGreater) {
 	auto acc = lua_get<uint32_t>(L, 3);
 	int32_t extraargs = lua_gettop(L) - 3;
 	card_vector cv(pduel->game_field->core.must_select_cards);
-	int32_t mcount = cv.size();
+	int32_t mcount = static_cast<int32_t>(cv.size());
 	const auto beginit = pduel->game_field->core.must_select_cards.begin();
 	const auto endit = pduel->game_field->core.must_select_cards.end();
 	for(auto& pcard : pgroup->container) {
@@ -525,7 +525,7 @@ LUA_FUNCTION(SelectWithSumGreater) {
 		pduel->game_field->core.select_cards.erase(it, pduel->game_field->core.select_cards.end());
 	}
 	card_vector cv(pduel->game_field->core.must_select_cards);
-	int32_t mcount = cv.size();
+	int32_t mcount = static_cast<int32_t>(cv.size());
 	cv.insert(cv.end(), pduel->game_field->core.select_cards.begin(), pduel->game_field->core.select_cards.end());
 	for(auto& pcard : cv)
 		pcard->sum_param = pduel->lua->get_operation_value(pcard, findex, extraargs);
@@ -883,7 +883,7 @@ LUA_FUNCTION_EXISTING(IsDeleted, is_deleted_object);
 void scriptlib::push_group_lib(lua_State* L) {
 	static constexpr auto grouplib = GET_LUA_FUNCTIONS_ARRAY();
 	static_assert(grouplib.back().name == nullptr, "");
-	lua_createtable(L, 0, grouplib.size() - 1);
+	lua_createtable(L, 0, static_cast<int>(grouplib.size() - 1));
 	luaL_setfuncs(L, grouplib.data(), 0);
 	lua_pushstring(L, "__index");
 	lua_pushvalue(L, -2);
